Change build logic from run time builds to build time builds
Build time mounts work using podman and somewhat recent docker, so this
is the natural way to build the image.

This allows for further merging with the GCC build, at least the commits
should be easily mergable. In the future there could be just one
Dockerfile which handles both cases